Refactor quantity types from int to float across broker and core modules
- Updated the `quantity` parameter in the `Broker` interface's `buy` and `sell` methods to accept float values instead of int.
- Changed the `_Holding` dataclass in `paper_broker.py` to use float for `quantity`.
- Adjusted the `TradeDecision`, `Position`, and `Fill` classes in `types.py` to reflect the change to float for `quantity`.
- Modified the `size_trade` function in `engine.py` to handle float quantities and allow fractional shares.
- Enhanced the `score_news` method in `AiProvider` and its implementations to accept market trends as an optional parameter.
- Updated the `MockAiProvider` to generate signals based on trends when no news is available.
- Improved the `run_one_cycle` function in `pipeline.py` to pre-compute market trends and pass them to the AI provider.
- Added debug information collection for news and AI processing in the pipeline.
- Ensured all relevant tests and type hints are updated to accommodate the new float type for quantities.

+++ src/aistock/broker/interface.py
@@ -7,11 +7,11 @@ from aistock.core.types import Fill, PortfolioSnapshot
 
 class Broker(ABC):
     @abstractmethod
-    def buy(self, symbol: str, quantity: int, price: float) -> Fill | None:
+    def buy(self, symbol: str, quantity: float, price: float) -> Fill | None:
         raise NotImplementedError
 
     @abstractmethod
-    def sell(self, symbol: str, quantity: int, price: float) -> Fill | None:
+    def sell(self, symbol: str, quantity: float, price: float) -> Fill | None:
         raise NotImplementedError
 
     @abstractmethod
